Simplify the imports by exposing the public API in __init__
